Moves MVRP join into SRP code.
The correct VLAN to join should be decided in SRP 
as this is where the domain VID is held. If 0 is 
passed as the VID from the application, this will
now join the domain VID and return the VID that 
was joined. The VID held by the AVB manager is then
up-to-date when queried by the API.